Detect if the pin value changes in the beginning
The value 0 is valid for pins so the initial value of the "previous"
values must be something different that will always not equal what can
be on the port. The maximum unsigned int value seems like a good
candidate since it's far away from any value representing a real input
or output port in the crosspoint.